requirements.txt
appdirs==1.4.0
attrs==16.3.0
Automat==0.5.0
cffi==1.9.1
constantly==15.1.0
cryptography==1.7.2
cssselect==1.0.1
Django==1.10.5
djangorestframework==3.5.4
enum34==1.1.6
idna==2.2
incremental==16.10.1
ipaddress==1.0.18
lxml==3.7.2
packaging==16.8
parsel==1.1.0
pkg-resources==0.0.0
psycopg2==2.6.2
pyasn1==0.2.2
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
queuelib==1.4.2
Scrapy==1.3.2
service-identity==16.0.0
six==1.10.0
SQLAlchemy==1.1.5
Twisted==17.1.0
urllib3==1.20
w3lib==1.17.0
zope.interface==4.3.3
